requirements.txt
asttokens==2.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
executing==0.8.3
humanize==4.0.0
icecream==2.1.2
idna==3.3
mdtable==0.2
pycparser==2.21
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
wrapt==1.14.0
yapf==0.32.0
